refactor(graph): split compiled/mod.rs into executor/state_api/routing
compiled/mod.rs had grown to ~2k lines mixing the superstep executor,
state-inspection/time-travel API, and step-routing logic in one impl
block. Split mechanically into executor.rs (run/resume/execute loop),
state_api.rs (get_state/update_state/fork_state), and routing.rs
(route/route_completed/interrupt-durability), keeping mod.rs for
shared helpers and the builder/configuration impl. No behavior change.
